IMU sensor library and cube visualization tool
These three folders allow us to visualize the data collected by the
IMU. It uses the library and the code from
http://bildr.org/2012/03/stable-orientation-digital-imu-6dof-arduino/ -
very grateful for the library. The code has been adapted to run on
Arduino DUE (instead of UNO for which the code and library are written
for).

To see the moving cube using a DUE and the IMU:

1. Connect the sensor’s data and clock to SDA20 and SCL21 respectively,
as well as the 5V and GND

2. Copy the FreeSixIMU_DUE (the library) into the libraries directory
in the local Arduino folder

3. Compile and download FreeSix_cube_DUE to the Arduino DUE

3. Load FreeSix_cube.pde using Processing

Processing is an application that can be downloaded from
https://processing.org/download/